# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"MedDesign MVP","description":"AI-powered pharmaceutical study design automation","version":"0.1.0"},"paths":{"/api/v1/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/search/start":{"post":{"tags":["Search"],"summary":"Start Search","description":"Start drug parameter search (async background task).\n\nReturns project_id immediately (202 Accepted pattern).","operationId":"start_search_api_v1_search_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchStartRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchStartResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/search/results/{project_id}":{"get":{"tags":["Search"],"summary":"Get Results","description":"Get search results and design parameters for a project.\nReturns current status and available data.","operationId":"get_results_api_v1_search_results__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResultsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/projects/{project_id}":{"get":{"tags":["Projects"],"summary":"Get Project Details","description":"Get full project details including design and regulatory status.","operationId":"get_project_details_api_v1_projects__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/{project_id}/generate":{"post":{"tags":["Reports"],"summary":"Generate Report","description":"Generate DOCX report for project.","operationId":"generate_report_api_v1_reports__project_id__generate_post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/{project_id}/download":{"get":{"tags":["Reports"],"summary":"Download Report","description":"Download generated report.","operationId":"download_report_api_v1_reports__project_id__download_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ParameterSchema":{"properties":{"parameter":{"type":"string","title":"Parameter"},"value":{"type":"string","title":"Value"},"unit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unit"},"source":{"type":"string","title":"Source"},"is_reliable":{"type":"boolean","title":"Is Reliable","default":true}},"type":"object","required":["parameter","value","source"],"title":"ParameterSchema","description":"Individual drug parameter from literature."},"SearchResultsResponse":{"properties":{"project_id":{"type":"string","title":"Project Id"},"status":{"type":"string","title":"Status"},"parameters":{"items":{"$ref":"#/components/schemas/ParameterSchema"},"type":"array","title":"Parameters"},"sources_count":{"type":"integer","title":"Sources Count"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["project_id","status","parameters","sources_count","created_at","updated_at"],"title":"SearchResultsResponse","description":"Response with search results and parameters."},"SearchStartRequest":{"properties":{"inn_en":{"type":"string","title":"Inn En","description":"Drug name in English"},"inn_ru":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Inn Ru","description":"Drug name in Russian"},"dosage":{"type":"string","title":"Dosage","description":"Drug dosage (e.g., '400mg')"},"form":{"type":"string","title":"Form","description":"Drug form (e.g., 'tablets')"}},"type":"object","required":["inn_en","dosage","form"],"title":"SearchStartRequest","description":"Request to start drug parameter search."},"SearchStartResponse":{"properties":{"project_id":{"type":"string","title":"Project Id"},"status":{"type":"string","title":"Status"},"message":{"type":"string","title":"Message"}},"type":"object","required":["project_id","status","message"],"title":"SearchStartResponse","description":"Response to search start request."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}